update profile 70%70
