+++ affsupplyroot/src/app/(main)/layout.tsx
@@ -1,0 +1,12 @@
+import Header from "@/layout/Header";
+import Footer from "@/layout/Footer";
+
+export default function MainLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <Header />
+      <main>{children}</main>
+      <Footer />
+    </>
+  );
+}
