+++ affsupplyroot/src/app/(main)/page.tsx
@@ -1,3 +1,4 @@
+import LayoutWrapper from "@/layout/LayoutWrapper";
 import HomePageView from "@/View/home";
 
 export default function Home() {
