+++ affsupplyroot/src/app/(supply)/profile/layout.tsx
@@ -1,0 +1,9 @@
+import Footer from "@/layout/Footer";
+import Header from "@/layout/Header";
+
+export default function SupplyNoLayout({ children }: { children: React.ReactNode }) {
+  return <>
+    <Header/>
+      {children}
+  </>;
+}
