+++ affsupplyroot/src/app/api/auth/logout/route.tsx
@@ -1,0 +1,7 @@
+import { NextResponse } from "next/server"
+
+export async function POST() {
+  const res = NextResponse.json({ message: "Logout success" })
+  res.cookies.set("token", "", { maxAge: 0 }) // clear cookie
+  return res
+}
