+++ affsupplyroot/src/app/authentication/layout.tsx
@@ -1,0 +1,3 @@
+export default function AuthenLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>
+}
