+++ affsupplyroot/src/layout/Header.tsx
@@ -17,6 +17,7 @@ import logo from "../../public/logo.png"
 import SearchCustomize from "@/components/search-customize"
 import { Button } from "@/components/ui/button"
 import { useRouter } from "next/navigation"
+import Link from "next/link"
 import useAuth from "@/hooks/useAuth"
 import { roleMenus } from "@/variable/menuHeader"
 
@@ -28,34 +29,38 @@ export default function Header() {
   const user = state.response
   const menus = user ? roleMenus[user?.user?.role] || [{ label: "Đăng xuất" }] : []
 
+  // 👉 gom logic xử lý click
+  const handleMenuClick = (item: any) => {
+    if (item.label === "Đăng xuất") {
+      handleLogout()
+      router.push("/authentication") // nếu muốn đẩy về login page
+    } else if (item.href) {
+      router.push(item.href)
+    }
+  }
 
-const renderMenuItems = () =>
-  menus.map((item, idx) => (
-    <DropdownMenuItem
-      key={idx}
-      onClick={() => {
-        if (item.label === "Đăng xuất") {
-          handleLogout()
-        } else if (item.href) {
-          router.push(item.href)
-        }
-      }}
-      className="flex items-center gap-3 px-3 py-2 text-yellow-400 hover:bg-green-800 hover:text-yellow-300 cursor-pointer rounded-md transition-colors duration-300"
-    >
-      {/* Icon */}
-      <span className="text-lg">{item.icon}</span>
-      {/* Label */}
-      <span className="font-medium">{item.label}</span>
-    </DropdownMenuItem>
-  ))
+  const renderMenuItems = () =>
+    menus.map((item, idx) => (
+      <DropdownMenuItem
+        key={idx}
+        onClick={() => handleMenuClick(item)}
+        className="flex items-center gap-3 px-3 py-2 text-yellow-400 hover:bg-green-800 hover:text-yellow-300 cursor-pointer rounded-md transition-colors duration-300"
+      >
+        <span className="text-lg">{item.icon}</span>
+        <span className="font-medium">{item.label}</span>
+      </DropdownMenuItem>
+    ))
 
   return (
     <header className="sticky top-0 z-30 w-full h-[5rem] bg-green-950 text-yellow-400 font-manuale lg:px-25 flex items-center shadow-md px-5 ">
+      {/* Mobile */}
       <div className="grid grid-cols-[auto_1fr_auto] items-center w-full md:hidden gap-2">
-        <motion.div whileHover={{ scale: 1.05 }} className="flex items-center gap-2 cursor-pointer">
-          <Image src={logo} alt="logo" width={40} height={40} className="rounded-md" />
-          <span className="text-lg font-bold hidden sm:inline">AFF supplyRoot</span>
-        </motion.div>
+        <Link href="/" className="flex items-center gap-2 cursor-pointer">
+          <motion.div whileHover={{ scale: 1.05 }} className="flex items-center gap-2">
+            <Image src={logo} alt="logo" width={40} height={40} className="rounded-md" />
+            <span className="text-lg font-bold hidden sm:inline">AFF supplyRoot</span>
+          </motion.div>
+        </Link>
 
         <div className="w-full max-w-[16rem] justify-self-center">
           <SearchCustomize />
@@ -72,33 +77,27 @@ const renderMenuItems = () =>
           ) : (
             <Sheet>
               <SheetTrigger asChild>
-              <motion.div whileHover={{ scale: 1.05 }} whileTap={{ scale: 0.95 }} className="flex items-center gap-2 cursor-pointer hover:text-green-400 transition">
-                <FaUserCircle className="w-6 h-6" />
-              </motion.div>
+                <motion.div whileHover={{ scale: 1.05 }} whileTap={{ scale: 0.95 }} className="flex items-center gap-2 cursor-pointer hover:text-green-400 transition">
+                  <FaUserCircle className="w-6 h-6" />
+                </motion.div>
               </SheetTrigger>
               <SheetContent side="right" className="bg-green-950 text-yellow-400 w-[260px] p-0">
-              <SheetHeader className="border-b border-yellow-700/30 px-6 pt-6 pb-3">
-                <SheetTitle className="text-yellow-300 text-lg font-bold">Tài khoản</SheetTitle>
-              </SheetHeader>
-              <div className="mt-4 flex flex-col gap-1 px-4">
-                {menus.map((item, idx) => (
-                <Button
-                  key={idx}
-                  variant="ghost"
-                  className="flex items-center gap-3 justify-start text-yellow-400 hover:bg-green-800 hover:text-yellow-300 rounded-lg px-3 py-2 transition-colors duration-200 text-base font-medium"
-                  onClick={() => {
-                  if (item.label === "Đăng xuất") {
-                    handleLogout()
-                  } else if (item.href) {
-                    router.push(item.href)
-                  }
-                  }}
-                >
-                  <span className="text-lg">{item.icon}</span>
-                  <span>{item.label}</span>
-                </Button>
-                ))}
-              </div>
+                <SheetHeader className="border-b border-yellow-700/30 px-6 pt-6 pb-3">
+                  <SheetTitle className="text-yellow-300 text-lg font-bold">Tài khoản</SheetTitle>
+                </SheetHeader>
+                <div className="mt-4 flex flex-col gap-1 px-4">
+                  {menus.map((item, idx) => (
+                    <Button
+                      key={idx}
+                      variant="ghost"
+                      className="flex items-center gap-3 justify-start text-yellow-400 hover:bg-green-800 hover:text-yellow-300 rounded-lg px-3 py-2 transition-colors duration-200 text-base font-medium"
+                      onClick={() => handleMenuClick(item)}
+                    >
+                      <span className="text-lg">{item.icon}</span>
+                      <span>{item.label}</span>
+                    </Button>
+                  ))}
+                </div>
               </SheetContent>
             </Sheet>
           )}
@@ -107,10 +106,12 @@ const renderMenuItems = () =>
 
       {/* Desktop */}
       <div className="hidden md:flex items-center justify-between w-full">
-        <motion.div whileHover={{ scale: 1.05 }} className="flex items-center gap-2 cursor-pointer">
-          <Image src={logo} alt="logo" width={40} height={40} className="rounded-md" />
-          <span className="text-lg font-bold hidden sm:inline">AFF supplyRoot</span>
-        </motion.div>
+        <Link href="/" className="flex items-center gap-2 cursor-pointer">
+          <motion.div whileHover={{ scale: 1.05 }} className="flex items-center gap-2">
+            <Image src={logo} alt="logo" width={40} height={40} className="rounded-md" />
+            <span className="text-lg font-bold hidden sm:inline">AFF supplyRoot</span>
+          </motion.div>
+        </Link>
 
         <div className="flex items-center gap-6">
           <div className="w-[300px]">
@@ -132,17 +133,17 @@ const renderMenuItems = () =>
                   {desktopOpen ? <ChevronUp className="w-4 h-4 text-yellow-primary" /> : <ChevronDown className="w-4 h-4 text-yellow-primary" />}
                 </motion.div>
               </DropdownMenuTrigger>
-                <DropdownMenuContent
-                  align="end"
-                  sideOffset={20}
-                  className="w-60 bg-green-950 text-yellow-400 rounded-xl shadow-2xl border border-yellow-700/40"
-                >
-                  <DropdownMenuLabel className="font-bold text-yellow-300 text-sm px-3 py-2">
-                    Tài khoản
-                  </DropdownMenuLabel>
-                  <DropdownMenuSeparator className="bg-yellow-700/30" />
-                  {renderMenuItems()}
-                </DropdownMenuContent>
+              <DropdownMenuContent
+                align="end"
+                sideOffset={20}
+                className="w-60 bg-green-950 text-yellow-400 rounded-xl shadow-2xl border border-yellow-700/40"
+              >
+                <DropdownMenuLabel className="font-bold text-yellow-300 text-sm px-3 py-2">
+                  Tài khoản
+                </DropdownMenuLabel>
+                <DropdownMenuSeparator className="bg-yellow-700/30" />
+                {renderMenuItems()}
+              </DropdownMenuContent>
             </DropdownMenu>
           )}
         </div>
