+++ affsupplyroot/src/variable/menuHeader.tsx
@@ -7,20 +7,20 @@ export const roleMenus: Record<
   customer: [
     { label: "Profile", href: "/profile", icon: <FaUser className="w-4 h-4" /> },
     { label: "Orders", href: "/orders", icon: <FaBox className="w-4 h-4" /> },
-    { label: "Đăng xuất", icon: <FaSignOutAlt className="w-4 h-4" /> },
+    { label: "Đăng xuất",href: "/", icon: <FaSignOutAlt className="w-4 h-4" /> },
   ],
   supply: [
     { label: "Profile", href: "/profile", icon: <FaUser className="w-4 h-4" /> },
     { label: "Orders", href: "/orders", icon: <FaBox className="w-4 h-4" /> },
-    { label: "Đăng xuất", icon: <FaSignOutAlt className="w-4 h-4" /> },
+    {  label: "Đăng xuất",href: "/", icon: <FaSignOutAlt className="w-4 h-4" /> },
   ],
   admin: [
     { label: "Dashboard", href: "/admin", icon: <FaTachometerAlt className="w-4 h-4" /> },
     { label: "Quản lý người dùng", href: "/admin/users", icon: <FaUsers className="w-4 h-4" /> },
-    { label: "Đăng xuất", icon: <FaSignOutAlt className="w-4 h-4" /> },
+    {  label: "Đăng xuất",href: "/", icon: <FaSignOutAlt className="w-4 h-4" /> },
   ],
   staff: [
     { label: "Dashboard", href: "/staff", icon: <FaTachometerAlt className="w-4 h-4" /> },
-    { label: "Đăng xuất", icon: <FaSignOutAlt className="w-4 h-4" /> },
+    {  label: "Đăng xuất",href: "/",  icon: <FaSignOutAlt className="w-4 h-4" /> },
   ],
 }
